Hls player docs (#1365)
* Added view for hls player

* Added play call

* Added hls player in android

* Added hls player support for android

* Removed pip hls player

* Completed HLS Player for android

* iOS HLS Player setup

* Release app

* released sample app version 1.4.107 (242) 🍀

* updated pubspec file

* corrected lint warnings

* added HMSHLSPlayerSDK dependency on iOS

* Added hls player in iOS

* Added comments

* updated packages

* fixed swift lint warnings

* updated changelog

* fixed dart warnings

* released sample app version 1.4.108 (243) 🍀

* updated packages

* Fixed hls player bugs (#1360)

* Released apps

* released sample app version 1.4.110 (245) 🍀

* updated packages

* released sample app version 1.4.111 (246) 🍀

* Added comments in code

* added comments in code

* Made acoustic echo canceller to false by default (#1368)

---------
