+++ lib/hmssdk_flutter.dart
@@ -21,6 +21,8 @@ export 'src/enum/hms_track_init_state.dart';
 export 'src/enum/hms_Quality_limitation_reason.dart';
 export 'src/enum/hms_simulcast_layer.dart';
 export 'src/enum/hms_audio_mode.dart';
+export 'src/enum/hms_hls_playback_state.dart';
+
 //EXCEPTIONS
 export 'src/exceptions/hms_exception.dart';
 export 'src/exceptions/hms_in_sufficient_data.dart';
@@ -91,9 +93,11 @@ export 'src/model/hms_log_list.dart';
 export 'src/model/hms_camera_controls.dart';
 export 'src/model/hms_session_store.dart';
 export 'src/model/hms_key_change_listener.dart';
-export 'src/model/hms_hls_playback_event_listener.dart';
-export 'src/model/hms_hls_player_controller.dart';
-export 'src/model/hms_hls_timed_metadata.dart';
+export 'src/model/hls_player/hms_hls_playback_event_listener.dart';
+export 'src/model/hls_player/hms_hls_player_controller.dart';
+export 'src/model/hls_player/hms_hls_player_stats.dart';
+export 'src/model/hls_player/hms_hls_cue.dart';
+export 'src/model/hls_player/hms_hls_timed_metadata.dart';
 
 //Views
 export 'src/ui/meeting/hms_video_view.dart';
