+++ lib/src/model/hls_player/hms_hls_cue.dart
@@ -1,9 +1,20 @@
+//Project imports
 import 'package:hmssdk_flutter/src/model/hms_date_extension.dart';
 
+///100ms HMSHLSCue
+///
+///[HMSHLSCue] is used to parse the timed metadata sent across the platforms.
 class HMSHLSCue {
+  // Unique id of the timed event
   final String? id;
+
+  // startDate of the timed event
   final DateTime startDate;
+
+  // endDate of the timed event
   final DateTime? endDate;
+
+  // String payload of the timed event
   final String? payload;
 
   HMSHLSCue({this.id, required this.startDate, this.endDate, this.payload});
