+++ lib/src/model/hls_player/hms_hls_timed_metadata.dart
