+++ lib/src/ui/meeting/hms_video_view.dart
@@ -83,9 +83,8 @@ class HMSVideoView extends StatelessWidget {
       {Key? key,
       required this.track,
       this.setMirror = false,
-      @Deprecated(
-          "matchParent is not longer necessary and will be removed in future version")
-      this.matchParent = true,
+      @Deprecated("matchParent is not longer necessary and will be removed in future version")
+          this.matchParent = true,
       this.scaleType = ScaleType.SCALE_ASPECT_FIT,
       this.disableAutoSimulcastLayerSelect = false})
       : super(key: key);
